Bug fix for duplicate groups in _POST
If a user who has permission to assign a new user to user groups selected any of the groups chosen in the plugin’s settings, you end up with multiple array values of the same group which threw an error. This solves that by making sure the values are unique.